requirements.txt
pwntools>=4.2.1